feat: add YAML configuration support for dataset generation
This change adds the ability to load dataset generation configurations from a YAML file. The `load_configs` function reads the YAML file and converts the data into a list of `ExpressionGenerationConfig` objects. This allows for more flexibility in configuring the dataset generation process, such as specifying the number of samples, operators, and other parameters.

The main function has also been added to allow running the dataset generation script from the command line, taking the configuration file path and output file path as arguments.